+++ packages/sdk/src/core/Chunk.js
@@ -1,4 +1,4 @@
-import { generateUid, each, getHash } from '@/shared'
+import { generateUid, each } from '@/shared'
 import { Status } from './constans.js'
 
 export default class Chunk {
@@ -12,44 +12,30 @@ export default class Chunk {
     this.chunkSize = this.opts.chunkSize
 
     this.fileHash = this.file.hash
-    this.hash = ''
     this.uid = generateUid('chunk_id')
     this.stardByte = this.chunkSize * index
     this.endByte = Math.min(this.stardByte + this.chunkSize, this.totalSize)
     this.size = this.endByte - this.stardByte
     this.chunkIndex = index
 
-    this.retries = this.opts.retries
+    this.maxRetries = this.opts.maxRetries
     this.xhr = null
     this.promise = null
     this.status = Status.Ready
     this.progress = 0
-    this.progressInFile = 0
+    this.fakeProgress = 0
     this.timer = null
   }
 
-  async prepareXhr() {
+  prepareXhr() {
     const data = new FormData()
     this.xhr.responseType = 'json'
     this.xhr.withCredentials = this.opts.withCredentials
     const blob = this.file.rawFile.slice(this.stardByte, this.endByte)
 
-    if (this.opts.hasChunkHash) {
-      if (!this.hash) {
-        try {
-          this.hash = await getHash(blob)
-        } catch {
-          //
-        }
-      }
-    }
-
     data.append(this.opts.name, blob)
     if (this.fileHash) {
-      data.append('fileHash', this.fileHash)
-    }
-    if (this.hash) {
-      data.append('hash', this.hash)
+      data.append('hash', this.fileHash)
     }
     data.append('id', this.uid)
     data.append('fileId', this.fileId)
@@ -61,7 +47,7 @@ export default class Chunk {
     each(this.opts.data, (val, key) => {
       data.append(key, val)
     })
-    this.xhr.open('POST', this.opts.target, true)
+    this.xhr.open('POST', this.opts.action, true)
 
     // 'setRequestHeader' on 'XMLHttpRequest': The object's state must be OPENED
     if ('setRequestHeader' in this.xhr) {
@@ -79,12 +65,10 @@ export default class Chunk {
 
       // eslint-disable-next-line no-unused-vars
       const failHandler = (e) => {
-        // console.log(e)
-        // this.progress = 0
-        // this.progressInFile = 0
-        // this.file.setProgress(this)
-        if (this.retries <= 0) {
-          this.file.removeChunkInUploadingQueue(this)
+        this.progress = 0
+        this.file.setProgress(this)
+        if (this.maxRetries <= 0) {
+          this.file.removeUploadingQueue(this)
           this.status = Status.Fail
           if (this.file.status === Status.Uploading) {
             this.file.uploadFile()
@@ -94,15 +78,15 @@ export default class Chunk {
         } else {
           this.timer = setTimeout(() => {
             this.send()
-            this.retries--
+            this.maxRetries--
             clearTimeout(this.timer)
           }, this.uploader.opts.retryInterval)
         }
       }
       const doneHandler = (e) => {
-        if (this.uploader.opts.successStatuses(this.xhr)) {
+        if (this.uploader.opts.requestSucceed(this.xhr)) {
           this.status = Status.Success
-          this.file.removeChunkInUploadingQueue(this)
+          this.file.removeUploadingQueue(this)
           if (this.file.status === Status.Uploading) {
             this.file.uploadFile()
           }
@@ -114,14 +98,15 @@ export default class Chunk {
 
       const progressHandler = (e) => {
         this.progress = Math.min(1, e.loaded / e.total)
-        this.progressInFile = this.progress * (this.size / this.file.size)
+        this.fakeProgress = Math.max(this.progress, this.fakeProgress)
+
         this.status = Status.Uploading
         this.file.status = Status.Uploading
         this.file.setProgress(this)
       }
 
       this.xhr = new XMLHttpRequest()
-      const data = await this.prepareXhr()
+      const data = this.prepareXhr()
       this.xhr.upload.addEventListener('progress', progressHandler)
       this.xhr.addEventListener('load', doneHandler, false)
       this.xhr.addEventListener('error', failHandler, false)