+++ packages/sdk/src/core/Uploader.js
@@ -1,35 +1,78 @@
-import { Event, extend, each } from '@/shared'
-import File from './File.js'
-import defaults from './defaults.js'
-import { Status, Events } from './constans.js'
+import File from '@/core/File'
+import Container from '@/core/Container'
+import { Event, extend, isFunction, isPromise } from '@/shared'
+import { defaults } from '@/core/defaults'
+import { Status, Events } from '@/core/constans'
 
-class Uploader extends Event {
+class Uploader {
   constructor(options) {
-    super()
+    this.container = new Container(this)
+    this.event = new Event()
 
-    this.opts = extend({}, defaults.options, options)
+    this.opts = extend({}, defaults, options)
     this.fileList = this.opts.fileList || []
     this.status = Status.Init
+    // 只注册一次
+    this.listenerFiles()
+  }
+
+  on(name, func) {
+    this.event.on(name, func)
+  }
+
+  emit(name, ...args) {
+    this.event.emit(name, ...args)
   }
 
   async addFiles(files) {
-    const originFiles = [...files]
+    const { limit, beforeAdd, attributes } = this.opts
 
-    if (originFiles.length + this.fileList.length > this.opts.limit) {
-      this.emit(Events.Exceed)
+    if (limit > 0) {
+      if (files.length + this.fileList.length > limit) {
+        this.emit(Events.Exceed, files, this.fileList)
+        return
+      }
+    }
+
+    let originFileList = [...files]
+
+    if (!this.opts.multiple) {
+      originFileList = originFileList.slice(0, 1)
+    }
+
+    if (originFileList.length === 0) {
       return
     }
 
-    const newFileList = originFiles.map((file) => new File(this, file))
-
-    this.status = Status.Ready
+    const newFileList = originFileList.map((file) => new File(this, file))
     this.fileList = [...this.fileList, ...newFileList]
-    this.emit(Events.FilesAdded, this.fileList)
 
-    await Promise.all(newFileList.map((file) => file.bootstrap()))
+    if (beforeAdd) {
+      if (isFunction(beforeAdd)) {
+        for (let i = 0; i < newFileList.length; i++) {
+          const file = newFileList[i]
+          const before = beforeAdd(file)
+          if (isPromise(before)) {
+            before.then(
+              () => {},
+              () => {
+                this.doRemove(file)
+              }
+            )
+          } else if (before !== false) {
+            //
+          } else {
+            this.doRemove(file)
+          }
+        }
+      }
+    }
+
+    this.emit(Events.FilesAdded, this.fileList)
+    this.status = Status.Ready
 
     if (this.opts.autoUpload) {
-      this.upload()
+      this.submit()
     }
   }
 
@@ -56,20 +99,32 @@ class Uploader extends Event {
         return
       }
     }
-    // every 在数组为空时会返回true
-    if (this.fileList.length) {
-      const allSuccess = this.fileList.every((file) => file.isSuccess())
-      if (allSuccess) {
-        this.emit(Events.AllFileSuccess, this.fileList)
-      }
-    }
   }
 
-  submit() {
+  async submit() {
+    const initedFileList = this.fileList.filter((file) => file.isInited())
+    // 第一个file ready之后就开始上传，避免多个ready状态的file同时上传
+    await Promise.race(initedFileList.map((file) => file.start()))
     this.upload()
   }
 
-  remove(file) {
+  listenerFiles() {
+    const emitAllSuccess = (fiel, fileList) => {
+      if (!fileList.length) {
+        return
+      }
+      const allSuccess = fileList.every((file) => file.isSuccess())
+      if (allSuccess) {
+        this.emit(Events.AllFileSuccess, this.fileList)
+        this.status = Status.Success
+      }
+    }
+
+    this.on(Events.FileSuccess, emitAllSuccess)
+    this.on(Events.FileRemove, emitAllSuccess)
+  }
+
+  doRemove(file) {
     if (!file) {
       this.clear()
       return
@@ -83,9 +138,36 @@ class Uploader extends Event {
     }
   }
 
+  clear() {
+    for (let i = 0; i < this.fileList.length; i++) {
+      const file = this.fileList[i]
+      file.remove()
+      this.emit(Events.FileRemove, file, this.fileList)
+    }
+    this.fileList = []
+    this.emit(Events.FileRemove, null, [])
+  }
+
+  remove(file) {
+    const { beforeRemove } = this.opts
+    if (!beforeRemove) {
+      this.doRemove(file)
+    } else if (isFunction(beforeRemove)) {
+      const before = beforeRemove(file)
+      if (isPromise(before)) {
+        before.then(() => {
+          this.doRemove(file)
+        })
+      } else if (before !== false) {
+        this.doRemove(file)
+      }
+    }
+  }
+
   retry(file) {
+    const index = this.fileList.indexOf(file)
     if (index > -1) {
-      file.retryUpload()
+      file.retry()
       this.upload()
     }
   }
@@ -104,78 +186,20 @@ class Uploader extends Event {
     }
   }
 
-  clear() {
-    for (let i = 0; i < this.fileList.length; i++) {
-      const file = this.fileList[i]
-      file.remove()
-      this.emit(Events.FileRemove, file, this.fileList)
-    }
-    this.fileList = []
-    this.emit(Events.FileRemove, null, [])
-  }
-
-  assignBrowse(domNode, attributes) {
-    let input
-    attributes = extend({}, defaults.attributes, attributes)
-    if (domNode.tagName === 'INPUT' && domNode.type === 'file') {
-      input = domNode
-    } else {
-      input = document.createElement('input')
-      input.setAttribute('type', 'file')
-
-      extend(input.style, {
-        visibility: 'hidden',
-        position: 'absolute',
-        width: '1px',
-        height: '1px'
-      })
-
-      each(attributes, (val, key) => {
-        input.setAttribute(key, val)
-      })
-      if (attributes.multiple) {
-        input.setAttribute('multiple', 'multiple')
-      } else {
-        input.removeAttribute('multiple')
-      }
-
-      domNode.appendChild(input)
-      domNode.addEventListener(
-        'click',
-        () => {
-          input.click()
-        },
-        false
-      )
-
-      input.addEventListener(
-        'change',
-        (e) => {
-          this.addFiles(e.target.files, e)
-          e.target.value = ''
-        },
-        false
-      )
-    }
+  assignBrowse(domNode, attributes = {}) {
+    const attrs = extend(
+      {},
+      {
+        accept: this.opts.accept,
+        multiple: this.opts.multiple
+      },
+      attributes
+    )
+    this.container.assignBrowse(domNode, attrs)
   }
 
   assignDrop(domNode) {
-    const preventEvent = (e) => {
-      e.preventDefault()
-    }
-    const handles = {
-      dragover: preventEvent,
-      dragenter: preventEvent,
-      dragleave: preventEvent,
-      drop: (e) => {
-        e.stopPropagation()
-        e.preventDefault()
-        this.addFiles(e.dataTransfer.files, e)
-      }
-    }
-    each(handles, (handler, name) => {
-      domNode.addEventListener(name, handler, false)
-    })
+    this.container.assignDrop(domNode)
   }
 }