+++ packages/sdk/src/shared/blob.js
@@ -1,0 +1,1 @@
+export const slice = File.prototype.slice || File.prototype.mozSlice || File.prototype.webkitSlice